HW-2 Rearrange project structure - Add domain package and put Movie class inside it - Rename MovieResponse to NetworkMovie - Add MovieDataTransferObjects file and put NetworkMovie class inside it
