+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/domain/Movie.kt
@@ -1,0 +1,11 @@
+package com.example.googryandroidarchitecturestudy.domain
+
+data class Movie(
+    val title: String,
+    val image: String,
+    val subtitle: String,
+    val pubDate: String,
+    val director: String,
+    val actor: String,
+    val userRating: String
+)
