+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/network/MovieApiService.kt
@@ -1,6 +1,5 @@
 package com.example.googryandroidarchitecturestudy.network
 
-import com.example.googryandroidarchitecturestudy.data.model.MovieResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -9,6 +8,6 @@ interface MovieApiService {
     @GET("/v1/search/movie.json")
     suspend fun searchMovies(
         @Query("query") query: String
-    ): MovieResponse
+    ): NetworkMovie
 
 }
