+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/network/MovieDataTransferObjects.kt
@@ -1,0 +1,11 @@
+package com.example.googryandroidarchitecturestudy.network
+
+import com.example.googryandroidarchitecturestudy.domain.Movie
+
+data class NetworkMovie(
+    val display: Int,
+    val items: List<Movie>,
+    val lastBuildDate: String,
+    val start: Int,
+    val total: Int
+)
